+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 - The bin symlinks are properly removed from `node_modules/.bin` when the corresponding dependencies
   are removed. This bug was a regression starting from 2.2.0. Integration tests added to prevent future regressions.
+- Transitive peer dependency requirements will be respected by `node-modules` linker in more edge cases
 
 ## 2.2.0